pyproject.toml
[project]
name = "taskshare-service"
version = "0.1.0"
description = "Task management API with async background processing (Celery + Redis), built with Django REST Framework and Docker."
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "django>=5.0,<6.0",
    "djangorestframework>=3.15,<3.16",
    "psycopg2-binary",
    "pydantic-settings",
    "drf-spectacular",
    "djangorestframework-simplejwt",
    "celery>=5.6.2",
    "redis>=7.3.0",
]

[dependency-groups]
dev = [
    "ruff>=0.14.6",
]
